QUAL-408 : Made buildqual script changes

+++ build/buildqual.sh
@@ -5,7 +5,7 @@ NMSDIR=~/tklabs-nms
 QUALDIR=~/qual
 MPSBUILDDIR=~/mps-builder
 
-BUILD="QUAL"
+BUILD=""
 TAG="NO"
 NMS="NO"
 BRANCH="dev/NMS"
@@ -48,30 +48,12 @@ titoutils() {
     tito build --rpm --tag=tklabs_utils-${UTILSVERSION} --offline
 }
 
-# Handle tito tag and build for nms
+# Handle tito tag and build for nms from qual tree
 titonms() {
-    if [ "$NMS" == "YES" ]; then
-        echo "Building nms RPMs from Thales Github repos! (/'-')/"
-
-        if [ ! -d ${NMSDIR} ]; then
-            cd
-            git clone https://github.com/mapcollab/tklabs-nms.git
-            cd ${NMSDIR}/
-        else
-            cd ${NMSDIR}/
-            git fetch origin master
-            git reset --hard FETCH_HEAD
-            git clean -df
-        fi
-
-        NMSVERSION=`cat .tito/packages/nms | cut -f 1 -d ' '`
-    else
-        echo "Building nms RPMs from QUAL tree! ( '-')"
-        cd ${QUALDIR}/src/nms
-        if [ "$TAG" == "YES" ]; then tito tag; fi
-        NMSVERSION=`cat ${QUALDIR}/.tito/packages/nms | cut -f 1 -d ' '`
-    fi
-
+    echo "Building nms RPMs from QUAL tree! ( '-')"
+    cd ${QUALDIR}/src/nms
+    if [ "$TAG" == "YES" ]; then tito tag; fi
+    NMSVERSION=`cat ${QUALDIR}/.tito/packages/nms | cut -f 1 -d ' '`
     tito build --rpm --tag=nms-${NMSVERSION} --offline
 }
 
@@ -188,14 +170,17 @@ set -e
 cd ${QUALDIR}/
 echo "Please use your own Git credentials to log in. \(^^\) \(^^)/ (/^^)/"
 
-git fetch --tags origin "$BRANCH"
-git checkout "$BRANCH"
-git reset --hard FETCH_HEAD
+if [ "$BRANCH" ]; then
+    git fetch --tags origin "$BRANCH"
+    git checkout "$BRANCH"
+    git reset --hard FETCH_HEAD
+fi
+
 git clean -df
 rm -rf /tmp/tito
 tito init
 titoutils
-titonms
+if [ "$NMS" == "YES" ]; then titonms; fi
 titoqual
 
 if [ "$TAG" == "YES" ]; then
